flask-api/requirements.txt
flask
argon2-cffi
requests
pyjwt